requirements.txt
Django>=1.3.1
South
Werkzeug
django-categories
django-debug-toolbar
django-extensions
lettuce
-e git+https://github.com/alex/django-taggit.git#egg=django-taggit
-e git+https://github.com/stard0g101/needle.git#egg=needle
-e git+https://github.com/dragoon/django-selenium.git#egg=django-selenium
-e git+https://github.com/stchris/untangle.git#egg=untangle
PIL
selenium
nose
unittest2

